Refactor of the sidesaddle class
The overriding principle here is that only elements that have a side class (which defaults to right placement) will go sidesaddle. No more extending the class elsewhere. This also introduces a middle step, called "notching," where things move out of the way a little but remain partially in the river.