Fixes several issues.
`.remove()` is not supported by IE11, so we wrote IE11-compliant code.

The addition of a seperate page broke the navigation (internal links) because it was id/# based. Now we check if we should be scrolling or jumping to a new page, and also checking for things like "open sidebar" buttons. If they are missing then the JS used to break.